move most of the ticker logic to tickerreply
